Added renderqueue; Added separate renderthread; Added editor pch; Added timer.

+++ engine/src/eepch.h
@@ -26,8 +26,10 @@
 #include <glm/gtc/matrix_transform.hpp>
 #include <glm/gtc/integer.hpp>
 
+#include "ref.h"
+
 #include "assrt.h"
 
 #include "util/log.h"
 
-#define NUMOF(x) ((sizeof(x)/sizeof(0[x])) / ((size_t)(!(sizeof(x) % sizeof(0[x])))))
+#define NUMOF(x) (uint32_t)(((sizeof(x)/sizeof(0[x])) / ((size_t)(!(sizeof(x) % sizeof(0[x]))))))
